Changed int-type to bigint.
Got a few truncation errors on integers.Changed data type for int to BigInteger.
RAM is cheap and most CPUs are 64 bit after all. So rather than yet anoter sizing-detector, like string just go for big-int.

+++ gentable.py
@@ -31,7 +31,7 @@ def load_to_table(fd,dbURL,sep=',', tabname=None,verbose=False):
 	if verbose:
 		print "Loading file:%s to table:%s"%(fd.name,tabname)
 	csv=csvparse.csvparse(fd,sep=sep,maxrows=10000,verbose=verbose)
-	typemap={datetime:DateTime, int:Integer, float:Float, unicode:String}
+	typemap={datetime:DateTime, int:BigInteger, float:Float, unicode:String}
 	cols=[]
